SplashPage(navigator) => components/SplashPage
