+++ android_ios/index.js
@@ -1,38 +1,19 @@
 import React, { Component } from 'react';
 import {
   Navigator,
-  View,
-  Text,
-  Image,
-  StyleSheet,
 } from 'react-native';
 
+import SplashPage         from './components/SplashPage'
 import LoginPage          from './components/LoginPage'
 import PickUpLocationPage from './components/PickUpLocationPage'
 
 export default class UberFooBarReactNativeFirebase extends Component {
-
-    SplashPage(navigator) {
-        const logo = require('../assets/img/logo.png')
-
-        setTimeout(() => {
-            navigator.replace({
-                id: 'LoginPageId',
-            });
-        }, 3000);
-
-        return (
-            <View style={SplashPageStyles.container}>
-                <Image source={logo}/>
-                <Text style={SplashPageStyles.splash}>SplashPage</Text>
-            </View>
-        );
-    }
-
     renderScene(route, navigator) {
         switch(route.id) {
             case 'SplashPageId':
-                return this.SplashPage(navigator);
+                return (
+                    <SplashPage navigator={navigator} />
+                );
 
             case 'LoginPageId':
                 return (
@@ -65,17 +46,3 @@ export default class UberFooBarReactNativeFirebase extends Component {
         );
     }
 }
-
-const SplashPageStyles = StyleSheet.create({
-    container: {
-        flex:1,
-        flexDirection:'row',
-        alignItems: 'center',
-        justifyContent: 'center',
-    },
-    splash: {
-        fontSize: 20,
-        textAlign: 'center',
-        margin: 10,
-    },
-});
